requirements.txt
fastapi
pydantic
web3
python-dotenv
uvicorn
requests
pandas